Docker version of LAMS
This Dockerfie together with the supporting scripts will spin up jenkins and lock it down, creates one user, creates a job and configures it. For proper operation it requires the user to run with at least a few specific environment variables:
HOST_IP - the ip of the machine hosting the container
GIT_TOKEN - a personal access token required by jenkins to execute api calls to set commit statuses
ACCOUNT_NAME - the git account name where the repo resides
REPO_NAME - the name of the git repo

Optional arguments are:
reporting, license_key and reporting_user which would help us track both efficiences and inefficiencies in the tool.

The user is also required to mount a set of .git ssh keys as well as a known_hosts file that verifies the git hostname which are needed to be able to pull and push to the repo being linted.